+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
         <p align="center">
           <img src="https://raw.githubusercontent.com/vinifmor/bauh-files/master/pictures/releases/0.9.22/missing_type_dep.png">
         </p>
+    
+- Web
+  - the Electron builds cache directory has been moved to the environment directory `~/.local/share/bauh/web/env/electron`
    
 
 ### Fixes